Refactoring installer/uninstaller review: https://github.com/forkcms/forkcms/pull/2233#pullrequestreview-61209838
