+++ src/Backend/Modules/Analytics/Installer/Installer.php
@@ -2,13 +2,13 @@
 
 namespace Backend\Modules\Analytics\Installer;
 
+use Backend\Core\Installer\AbstractModuleInstaller;
 use Backend\Core\Installer\InstallerInterface;
-use Backend\Core\Installer\ModuleInstaller;
 
 /**
  * Installer for the analytics module
  */
-class Installer extends ModuleInstaller implements InstallerInterface
+class Installer extends AbstractModuleInstaller implements InstallerInterface
 {
     public function install(): void
     {
