+++ src/Backend/Modules/Blog/Installer/Installer.php
@@ -9,14 +9,14 @@ namespace Backend\Modules\Blog\Installer;
  * file that was distributed with this source code.
  */
 
+use Backend\Core\Installer\AbstractModuleInstaller;
 use Backend\Core\Installer\InstallerInterface;
-use Backend\Core\Installer\ModuleInstaller;
 use Common\ModuleExtraType;
 
 /**
  * Installer for the blog module
  */
-class Installer extends ModuleInstaller implements InstallerInterface
+class Installer extends AbstractModuleInstaller implements InstallerInterface
 {
     /** @var int - Default category id */
     private $defaultCategoryId;
