+++ src/Backend/Modules/Mailmotor/Installer/Installer.php
@@ -9,14 +9,14 @@ namespace Backend\Modules\Mailmotor\Installer;
  * file that was distributed with this source code.
  */
 
+use Backend\Core\Installer\AbstractModuleInstaller;
 use Backend\Core\Installer\InstallerInterface;
-use Backend\Core\Installer\ModuleInstaller;
 use Common\ModuleExtraType;
 
 /**
  * Installer for the Mailmotor module
  */
-class Installer extends ModuleInstaller implements InstallerInterface
+class Installer extends AbstractModuleInstaller implements InstallerInterface
 {
     /** @var int */
     private $subscribeBlockId;
