+++ src/Backend/Modules/Users/Installer/Installer.php
@@ -9,8 +9,8 @@ namespace Backend\Modules\Users\Installer;
  * file that was distributed with this source code.
  */
 
+use Backend\Core\Installer\AbstractModuleInstaller;
 use Backend\Core\Installer\InstallerInterface;
-use Backend\Core\Installer\ModuleInstaller;
 use Backend\Modules\Groups\Engine\Model as BackendGroupsModel;
 use Backend\Modules\Profiles\Engine\Model as BackendProfilesModel;
 use Backend\Modules\Users\Engine\Model as BackendUsersModel;
@@ -19,17 +19,8 @@ use Symfony\Component\Filesystem\Filesystem;
 /**
  * Installer for the users module
  */
-class Installer extends ModuleInstaller implements InstallerInterface
+class Installer extends AbstractModuleInstaller implements InstallerInterface
 {
-    /**
-     * @var array
-     */
-    protected $promptVariables = [
-        'default_interface_language' => 'Default interface language (ex. ru, en)',
-        'email'                      => 'Email of admin',
-        'password'                   => 'Password of admin',
-    ];
-
     public function install(): void
     {
         $this->addModule('Users');
