+++ src/Console/Exceptions/UninstallerInterfaceException.php
@@ -9,12 +9,13 @@ namespace Console\Exceptions;
  * file that was distributed with this source code.
  */
 
+use Throwable;
+
 /**
  * Uninstaller interface exception class
  */
 class UninstallerInterfaceException extends \Exception
 {
-
     /**
      * UninstallerInterfaceException constructor.
      * @param string $module
@@ -24,14 +25,14 @@ class UninstallerInterfaceException extends \Exception
      * @param \Throwable|null $previous
      */
     public function __construct(
-        $module,
-        $uninstallerFile,
-        $message = '',
-        $code = 0,
-        \Throwable $previous = null
+        string $module,
+        string $uninstallerFile,
+        string $message = '',
+        int $code = 0,
+        Throwable $previous = null
     ) {
         $message = '' === $message
-            ? sprintf('Uninstaller class for module `%s` (%s) not implement UninstallerInterface.', $module, $uninstallerFile)
+            ? sprintf('Uninstaller class for module `%s` (%s) does not implement the UninstallerInterface.', $module, $uninstallerFile)
             : sprintf($message, $module, $uninstallerFile);
 
         parent::__construct($message, $code, $previous);
